Fixed RebootOptions / Unclassified Logic
If you select unclassified as a classification this essentially sets the options on the schedule to RebootOnly.  No other classifications will be set and no KBIDs will be set due to the fact that the schedule is meant to reboot only.  Reboot Only or Unclassified will never install updates, it will set the schedule to always reboot.